Add tests for utility.h and sysincludes.h macros
BUG=chromium-os:17564
TEST=make && make runtests

Reviewed-on: http://gerrit.chromium.org/gerrit/6608
Tested-by: Randall Spangler <rspangler@chromium.org>

+++ firmware/include/utility.h
@@ -46,7 +46,7 @@
 /* Combine [msw] and [lsw] uint16s to a uint32_t with its [msw] and
  * [lsw] forming the most and least signficant 16-bit words.
  */
-#define CombineUint16Pair(msw,lsw) (((msw) << 16) |     \
+#define CombineUint16Pair(msw,lsw) (((uint32_t)(msw) << 16) |   \
                                     (((lsw)) & 0xFFFF))
 /* Return the minimum of (a) or (b). */
 #define Min(a, b) (((a) < (b)) ? (a) : (b))
